use feature hooks class for managing the hooks
